UX/backend/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
pandas
numpy
scipy
sklearn
xgboost==1.4.2
matplotlib
seaborn
tqdm
click==7.1.1
# TODO - add git link to predictive OS package